requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
python-dotenv==1.0.1
google-generativeai==0.3.2
pip install Flask python-dotenv google-generativeai
pip install flask flask-session python-dotenv

##Caso tenha um ambiente e queria baixar mais rápido
pip install -r requirements.txt
